refactor: Refactor Profile routing
