--- packages/client/src/pages/Profile/Layout/Layout.tsx
@@ -1,9 +1,0 @@
-import { FC, PropsWithChildren } from 'react';
-
-export const Layout: FC<PropsWithChildren> = ({ children }) => {
-  return (
-    <div className="min-h-[calc(100vh-1.75rem)] max-w-4xl my-0 mx-auto p-8 bg-white relative border border-gray-300">
-      {children}
-    </div>
-  );
-};
